+++ include/exiv2/jp2image.hpp
@@ -50,7 +50,7 @@ class EXIV2API Jp2Image : public Image {
           not valid (does not look like data of the specific image type).
     @warning This function is not thread safe and intended for exiv2 -pS for debugging.
    */
-  void printStructure(std::ostream& out, PrintStructureOption option, int depth) override;
+  void printStructure(std::ostream& out, PrintStructureOption option, size_t depth) override;
 
   /*!
     @brief Todo: Not supported yet(?). Calling this function will throw
